Update footer.tpl
The prior version threw errors with PHP 7.0.33 and made the whole photos in an album not being shown.
I propose to use this syntax https://www.smarty.net/docs/en/tips.default.var.handling.tpl - it works for my setup with php 7.0.33, but I cannot test with  PHP8 and I am not a trained developer. I hope this helps.